Add SessionPolicy
